fix(Stream): fix implicit `ssize_t` cast issues with some compilers

+++ include/zenkit/Stream.hh
@@ -217,7 +217,7 @@ namespace zenkit {
 
 			cb(w);
 
-			auto len = static_cast<ssize_t>(w->tell()) - size_off - sizeof(uint32_t);
+			auto len = static_cast<ssize_t>(w->tell()) - size_off - static_cast<ssize_t>(sizeof(uint32_t));
 			w->seek(size_off, Whence::BEG);
 			w->write_uint(static_cast<uint32_t>(len));
 			w->seek(len, Whence::CUR);
